fix: bug of endless redirection
Create a home page that is different from the login page. When already
authenticated, stay on it.

Previously, when accessing the home page, it redirected to the login
page, but if it was already logged, it would redirect to the home page
again, entering in an endless loop with the token stored.

Signed-off-by: Lucas Cardoso <lucascardoso910@gmail.com>